add task

+++ src/main/java/com/hk/todotodo/dto/GoalDto.java
@@ -12,8 +12,8 @@ import java.time.LocalDate;
 @AllArgsConstructor
 // @ToString
 public class GoalDto {
-    private Long unq_id;
-    private Long category_id;
+    private int unq_id;
+    private int category_id;
     private LocalDate start_date;
     private LocalDate end_date;
 }
